+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.11.7
 
 - Added `numericSlider`, `enumSelector`, `textCommand` and `rawValue` values to `RenderWidget` enum.
+- Added `receptionAssetId` (`String?`) field to `AtsPurchaseOrder` model.
 
 ## 3.11.6